# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Browser4 WebDriver-Compatible API
  description: |
    A WebDriver-compatible HTTP API with selector-first extensions for browser automation.
    This API provides session management, navigation, element interaction, and event handling
    capabilities.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8182
    description: Local development server
...
paths:
...
  /session/{sessionId}:
    get:
      tags:
        - session
      summary: Get session details
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetails'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - session
      summary: Delete a session
      operationId: deleteSession
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /session/{sessionId}/event-configs:
    post:
      tags:
        - events
      summary: Create event configuration
      operationId: createEventConfig
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventConfig'
      responses:
        '200':
          description: Event config created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventConfigResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - events
      summary: Get event configurations
      operationId: getEventConfigs
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Event configurations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventConfigsResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/events:
    get:
      tags:
        - events
      summary: Get events
      operationId: getEvents
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Events list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/events/subscribe:
    post:
      tags:
        - events
      summary: Subscribe to events
      operationId: subscribeToEvents
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscribeRequest'
      responses:
        '200':
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/events/stream:
    get:
      tags:
        - events
      summary: Stream session events (Server-Sent Events)
      operationId: streamEvents
      parameters:
        - $ref: '#/components/parameters/sessionId'
        - name: subscriptionId
          in: query
          required: false
          description: If provided, stream only event types included in the subscription.
          schema:
            type: string
        - name: eventTypes
          in: query
          required: false
          description: Optional event type filter (repeatable query param).
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        '200':
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== PerceptiveAgent Endpoints ==========
...
components:
  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      description: Session identifier
...
  schemas:
...
    SessionDetails:
      type: object
      properties:
        value:
          type: object
          properties:
            sessionId:
              type: string
            url:
              type: string
            status:
              type: string
              enum: [active, paused, stopped]
            capabilities:
              type: object
              additionalProperties: true
    WebDriverResponse:
      type: object
      properties:
        value:
          description: Response value (can be null for success)
    ErrorResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            error:
              type: string
            message:
              type: string
            stacktrace:
              type: string
...
    EventConfig:
      type: object
      required:
        - eventType
      properties:
        eventType:
          type: string
          description: Event type to listen for
        selector:
          type: string
          description: Optional CSS selector to scope events
        enabled:
          type: boolean
          default: true
    EventConfigResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            configId:
              type: string
            eventType:
              type: string
            enabled:
              type: boolean
    EventConfigsResponse:
      type: object
      properties:
        value:
          type: array
          items:
            type: object
            properties:
              configId:
                type: string
              eventType:
                type: string
              enabled:
                type: boolean
    Event:
      type: object
      properties:
        eventId:
          type: string
        eventType:
          type: string
        timestamp:
          type: integer
          format: int64
        data:
          type: object
          additionalProperties: true
    EventsResponse:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/Event'
    SubscribeRequest:
      type: object
      required:
        - eventTypes
      properties:
        eventTypes:
          type: array
          items:
            type: string
          description: Event types to subscribe to
    SubscriptionResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            subscriptionId:
              type: string
            eventTypes:
              type: array
              items:
                type: string

    # ========== PerceptiveAgent Schemas ==========
...